Completeness first outsourcing
